Replace postgres with sqlite
So I want a portable application that I can "install" on my machine, without having to host it on some server or have to keep docker running. One move for this is introducing sqlite, so that I can easily use a database without the server and backup the database file.