Refactor Canvas related functionality to Web and new context
I already notice some of the benefits of being way more flexible when not using DDD principles.

Also, what Jimmy Bogard said regarding having a vertical slice in a page file is immediately noticable. There is no apparent disadvantage of having the code not in usecase classes. On the contrary, the test naming becomes obvious AND I can test the web project as well.